+++ changes.md
@@ -11,7 +11,8 @@ This project loosely follows Keep a Changelog and uses Semantic Versioning.
 ### Added
 
 - Find button discord RPC now links to YouTube Music search
-- Added irclib fallback for lyrics
+- Irclib fallback for lyrics
+- Albums screen
 - [Android] Storage permissions handler
 - [Android] URI path resolution
 - [Android] Folder manager access on mobile